requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
jinja2==3.1.4
python-dotenv==1.0.1
twilio==9.3.2
stripe==10.8.0
